make dist now support multi-locale: one package per locale.

+++ distribute/distribute.sh
@@ -1,14 +1,14 @@
 #!/bin/bash
-echo "Did you run 'make test' (Y/n)?"
-read answer
-test "$answer" = "n" && exit 1
-
 
 
 echo "Determining Version:"
 VERSION=$(sed -ne 's@^v\([a-z0-9\.]*\).*@\1@pi' ../paths2openscad.inx)
 echo "Version is: \"$VERSION\""
 name=inkscape-paths2openscad
+
+inx_basenames="paths2openscad"                  # we pack all inx_basenames in each locale
+inx_locales=".inx -de.inx"                      # we generate one package per inx_locale
+
 if [ -d $name ]
 then
 	echo "Removing leftover files"
@@ -22,15 +22,27 @@ cp -a ../make-includes $name/
 cp ../*.py ../*.inx ../Makefile $name/
 
 
+
+echo "Did you run 'make test' (Y/n)?"
+read answer
+test "$answer" = "n" && exit 1
 echo "****************************************************************"
 echo "Ubuntu Version: For Building you must have checkinstall and dpkg"
 echo "Build Ubuntu Version (Y/n)?"
 read answer
 if [ "$answer" != "n" ]
 then
+  rm -f out/*.deb
   mkdir -p deb/files
   cp -a $name/* deb/files
-  (cd deb && sh ./dist.sh $name $VERSION)
+  for inx in $inx_locales; do
+    rm -f deb/files/*.inx
+    for base in $inx_basenames; do
+      cp $name/$base$inx deb/files
+    done
+    pkgname=$(basename $name$inx .inx)
+    (cd deb && sh ./dist.sh $pkgname $VERSION)
+  done
 fi
